Refactor player machine types
